pyproject.toml
[project]
name = "ailab"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "email-validator>=2.2.0",
    "eventlet>=0.39.1",
    "flask>=3.1.0",
    "flask-login>=0.6.3",
    "flask-migrate>=4.1.0",
    "flask-socketio>=5.5.1",
    "flask-sqlalchemy>=3.1.1",
    "flask-wtf>=1.2.2",
    "qrcode[pil]>=8.0",
]
